splittin filters too
