+++ hat/assets/js/apps/Iaso/domains/registry/components/selectedOrgUnit/index.tsx
@@ -1,16 +1,8 @@
 import { Box, Divider, Paper } from '@mui/material';
 import { makeStyles } from '@mui/styles';
 import { LoadingSpinner, commonStyles } from 'bluesquare-components';
-import moment from 'moment';
-import React, { FunctionComponent, useCallback, useMemo } from 'react';
-import { useDispatch } from 'react-redux';
+import React, { FunctionComponent, useMemo } from 'react';
 
-import { baseUrls } from '../../../../constants/urls';
-import MESSAGES from '../../messages';
-
-import { redirectToReplace } from '../../../../routing/actions';
-
-import InputComponent from '../../../../components/forms/InputComponent';
 import InstanceFileContent from '../../../instances/components/InstanceFileContent';
 
 import {
@@ -22,6 +14,7 @@ import { OrgUnit } from '../../../orgUnits/types/orgUnit';
 import { HEIGHT } from '../../config';
 import { RegistryParams } from '../../types';
 import { EmptyInstances } from './EmptyInstances';
+import { Filters } from './Filters';
 import { InstanceTitle } from './InstanceTitle';
 import { OrgUnitTitle } from './OrgUnitTitle';
 
@@ -49,7 +42,6 @@ export const SelectedOrgUnit: FunctionComponent<Props> = ({
     isFetching: isFetchingOrgUnit,
 }) => {
     const classes: Record<string, string> = useStyles();
-    const dispatch = useDispatch();
 
     // selected instance should be:
     // submission id from params  OR reference instance OR first submission of the possible ones OR undefined
@@ -62,25 +54,7 @@ export const SelectedOrgUnit: FunctionComponent<Props> = ({
     const { data: currentInstance, isFetching: isFetchingCurrentInstance } =
         useGetInstance(currentInstanceId);
     const { data: instances, isFetching } = useGetOrgUnitInstances(orgUnit?.id);
-    const instancesOptions = useMemo(() => {
-        return (instances || []).map(instance => ({
-            label: `${instance.form_name} (${moment
-                .unix(instance.created_at)
-                .format('LTS')})`,
-            value: instance.id,
-        }));
-    }, [instances]);
 
-    const handleChange = useCallback(
-        (_, submissionId) => {
-            const newParams: RegistryParams = {
-                ...params,
-                submissionId,
-            };
-            dispatch(redirectToReplace(baseUrls.registry, newParams));
-        },
-        [params, dispatch],
-    );
     if (!orgUnit) {
         return null;
     }
@@ -99,31 +73,12 @@ export const SelectedOrgUnit: FunctionComponent<Props> = ({
                 <Divider />
                 {instances && instances?.length === 0 && <EmptyInstances />}
                 {instances && instances?.length > 0 && (
-                    <Box
-                        width="100%"
-                        display="flex"
-                        justifyContent="flex-end"
-                        mb={2}
-                        px={2}
-                    >
-                        <Box width="50%">
-                            <InputComponent
-                                type="select"
-                                disabled={
-                                    isFetching || instancesOptions.length <= 1
-                                }
-                                keyValue="instance"
-                                onChange={handleChange}
-                                value={
-                                    isFetching ? undefined : currentInstanceId
-                                }
-                                label={MESSAGES.submission}
-                                options={instancesOptions}
-                                loading={isFetching}
-                                clearable={false}
-                            />
-                        </Box>
-                    </Box>
+                    <Filters
+                        orgUnit={orgUnit}
+                        params={params}
+                        instances={instances}
+                        isFetching={isFetching}
+                    />
                 )}
 
                 {instances && instances?.length > 0 && currentInstance && (
